Project is fully work.

+++ src/com/d9nich/generator/Generatable.java
@@ -5,5 +5,5 @@ public interface Generatable {
      * @param fileName name of resulting file
      * @param dataSize how many integers we should write to file
      */
-    void generate(String fileName, long dataSize);
+    void generate(String fileName, int dataSize);
 }
